# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "maxLength": 254,
                    "format": "email",
                    "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 1024
                  }
                },
                "required": [
                  "email",
                  "password"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signed in. Session cookie set.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "identity": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                        },
                        "email": {
                          "type": "string"
                        },
                        "displayName": {
                          "type": "string"
                        },
                        "mfaEnabled": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "id",
                        "email",
                        "displayName",
                        "mfaEnabled"
                      ],
                      "additionalProperties": false
                    },
                    "memberships": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "organizationId": {
                            "type": "string",
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                          },
                          "organizationName": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "organizationId",
                          "organizationName"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "activeOrganizationId": {
                      "nullable": true,
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "grants": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "permission": {
                            "type": "string"
                          },
                          "scopeType": {
                            "type": "string",
                            "enum": [
                              "ORGANIZATION",
                              "PROPERTY"
                            ]
                          },
                          "propertyId": {
                            "nullable": true,
                            "type": "string",
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                          }
                        },
                        "required": [
                          "permission",
                          "scopeType",
                          "propertyId"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "csrfToken": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "identity",
                    "memberships",
                    "activeOrganizationId",
                    "grants",
                    "csrfToken"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "VALIDATION_FAILED",
                        "UNAUTHENTICATED",
                        "INVALID_CREDENTIALS",
                        "ACCOUNT_LOCKED",
                        "CSRF_FAILED",
                        "NO_ACTIVE_ORGANIZATION",
                        "FORBIDDEN",
                        "NOT_FOUND",
                        "CONFLICT",
                        "VERSION_CONFLICT",
                        "PRECONDITION_REQUIRED",
                        "IDEMPOTENCY_KEY_REUSED",
                        "IDEMPOTENCY_IN_PROGRESS",
                        "RATE_LIMITED",
                        "INTERNAL_ERROR"
                      ]
                    },
                    "detail": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "path": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "path",
                          "message"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "code"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/switch-organization": {
      "post": {
        "operationId": "AuthController_switchOrganization",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "organizationId": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  }
                },
                "required": [
                  "organizationId"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "identity": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                        },
                        "email": {
                          "type": "string"
                        },
                        "displayName": {
                          "type": "string"
                        },
                        "mfaEnabled": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "id",
                        "email",
                        "displayName",
                        "mfaEnabled"
                      ],
                      "additionalProperties": false
                    },
                    "memberships": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "organizationId": {
                            "type": "string",
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                          },
                          "organizationName": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "organizationId",
                          "organizationName"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "activeOrganizationId": {
                      "nullable": true,
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "grants": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "permission": {
                            "type": "string"
                          },
                          "scopeType": {
                            "type": "string",
                            "enum": [
                              "ORGANIZATION",
                              "PROPERTY"
                            ]
                          },
                          "propertyId": {
                            "nullable": true,
                            "type": "string",
                            "format": "uuid",
                            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                          }
                        },
                        "required": [
                          "permission",
                          "scopeType",
                          "propertyId"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "csrfToken": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "identity",
                    "memberships",
                    "activeOrganizationId",
                    "grants",
                    "csrfToken"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Hotel Platform API",
    "description": "Multi-tenant hotel management platform. Errors use application/problem+json (RFC 9457).",
    "version": "v1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "hotel_sid"
      }
    },
...
  }
}